enhanced_requirements.txt
# Core ML/Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.0.0
gender-guesser>=0.4.0

# Text Processing
regex>=2021.10.23

# Web/API
flask>=2.0.0
requests>=2.26.0

# Visualization (optional)
matplotlib>=3.4.0
seaborn>=0.11.0

# Development
tqdm>=4.62.0
python-dotenv>=0.19.0
